requirements.txt
djangorestframework==3.12.4
Pillow==8.2.0
requests==2.25.1
tqdm==4.60.0
